+++ deployment/terraform/batch.tf
@@ -69,6 +69,8 @@ data "template_file" "default_job_definition" {
     batch_job_queue_name             = "queue${local.short}Default"
     batch_job_def_name               = "job${local.short}Default"
     log_group_name                   = "log${local.short}Batch"
+    cache_host                       = aws_route53_record.cache.name
+    cache_port                       = var.ec_memcached_port
   }
 }
 
@@ -155,6 +157,8 @@ data "template_file" "notifications_job_definition" {
     batch_job_queue_name             = "queue${local.short}Notifications"
     batch_job_def_name               = "job${local.short}Notifications"
     log_group_name                   = "log${local.short}Batch"
+    cache_host                       = aws_route53_record.cache.name
+    cache_port                       = var.ec_memcached_port
   }
 }